backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pandas==2.2.3
python-multipart==0.0.17
python-dotenv==1.0.1
openai==1.57.4
google-generativeai==0.8.3
sqlalchemy==2.0.36
python-pptx==1.0.2
reportlab==4.2.5
matplotlib==3.9.3
seaborn==0.13.2
plotly==5.24.1
kaleido==0.2.1
openpyxl==3.1.5
aiofiles==24.1.0
jinja2==3.1.4
pydantic==2.10.3
